Translation script to write csv into xml
