--- translated_by_canada/ActivityDateType.csv
@@ -1,5 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Planned start,Début prévu,"The date on which the activity is planned to start, for example the date of the first planned disbursement or when physical activity starts.","La date à laquelle on prévoit que l'activité commencera, par ex. la date du premier décaissement prévu ou la date où les activités concrètes commenceront.",,
-2,Actual start,Début réel,"The actual date the activity starts, for example the date of the first disbursement or when physical activity starts.","La date à laquelle l'activité commence réellement, par ex. la date du premier décaissement ou la date où les activités concrètes commencent.",,
-3,Planned End,Fin prévue,"The date on which the activity is planned to end, for example the date of the last planned disbursement or when physical activity is complete.","La date à laquelle on prévoit que l'activité prendra fin, par ex. la date du dernier décaissement prévu ou la date où les activités concrètes se termineront.",,
-4,Actual end,Fin réelle,"The actual date the activity ends, for example the date of the last disbursement or when physical activity is complete.","La date à laquelle l'activité prend réellement fin, par ex. la date du dernier décaissement prévu ou la date où les activités concrètes se terminent.",,
