--- translated_by_canada/AidTypeVocabulary.csv
@@ -1,2 +1,0 @@
-code,name,name (FR),description,category,url
-1,OECD DAC,CAD OCDE,,,
