--- translated_by_canada/BudgetIdentifierVocabulary.csv
@@ -1,6 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,IATI,IITA,The budget identifier reported uses IATI budget identifier categories,L'identifiant budgétaire déclaré se sert des catégories d'identification budgétaire de l'IITA,,
-2,Country Chart of Accounts,Plan comptable national,The budget identifier reported corresponds to the recipient country chart of accounts,L'identifiant budgétaire déclaré correspond au plan comptable national du pays bénéficiaire,,
-3,Other Country System,Autre système national,The budget identifier reported corresponds to a recipient country system other than the chart of accounts,L'identifiant budgétaire déclaré correspond à un système du pays bénéficiaire autre que le plan comptable,,
-4,Reporting Organisation,Organisation déclarante,The budget identifier reported corresponds to categories that are specific to the reporting organisation,L'identifiant budgétaire déclaré correspond à des catégories particulières à l'organisation déclarante,,
-5,Other,Autre,"The budget identifier reported uses a different vocabulary, not specified in the codelist",L'identifiant budgétaire déclaré emploie un vocabulaire différent ne faisant pas partie de la liste des codes,,
