--- translated_by_canada/BudgetNotProvided.csv
@@ -1,4 +1,0 @@
-code,name,name (FR),description,category,url
-1,Commercial Restrictions,Restrictions commerciales,,,
-2,Legal Restrictions,Restrictions légales,,,
-3,Rapid Onset Emergency,Urgence à développement rapide,,,
