--- translated_by_canada/BudgetStatus.csv
@@ -1,3 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Indicative,Indicatif,A non-binding estimate for the described budget.,Une estimation non contraignante du budget décrit.,,
-2,Committed,Engagé,A binding agreement for the described budget.,Une entente contraignante au sujet du budget décrit.,,
