--- translated_by_canada/BudgetType.csv
@@ -1,3 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Original,Original,The original budget allocated to the activity,Le budget initialement affecté à l'activité.,,
-2,Revised,Révisé,The updated budget for an activity,Le budget mis à jour pour une activité.,,
