--- translated_by_canada/CRSAddOtherFlags.csv
@@ -1,5 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Free standing technical cooperation,Coopération technique pure,Free standing technical cooperation (Col 24 in CRS++ Reporting format),Coopération technique pure (col 24 du gabarit CRS++),,
-2,Programme-based approach,Approche-programme,Programme-based approach (Col 25 in CRS++ Reporting format),Approche-programme (col 25 du gabarit CRS++),,
-3,Investment project,Projet d'investissement,Investment project (Col 26 in CRS++ Reporting format),Projet d'investissement (col 26 du gabarit CRS++),,
-4,Associated financing,Financement associé,Associated financing (Col 27 in CRS++ Reporting format),Financement associé (col 27 du gabarit CRS++),,
