--- translated_by_canada/ContactType.csv
@@ -1,5 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,General Enquiries,Renseignements généraux,General Enquiries,Renseignements généraux,,
-2,Project Management,Gestion de projet,Project Management,Gestion de projet,,
-3,Financial Management,Gestion financière,Financial Management,Gestion financière,,
-4,Communications,Communications,Communications,Communications,,
