--- translated_by_canada/DescriptionType.csv
@@ -1,5 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,General,Général,"Unstructured, long description of the activity",Description longue et non structurée de l'activité,,
-2,Objectives,Objectifs,"Specific objectives for the activity, e.g. taken from logical framework","Objectifs précis pour l'activité, p. ex. tirés d'un cadre logique",,
-3,Target Groups,Groupes cibles,Details of groups that are intended to benefit from the activity,Détails sur les groupes qui devraient profiter de l'activité,,
-4,Other,Autre,For miscellaneous use. A further classification or breakdown may be included in the narrative,Pour utilisation diverse. Une classification ou une ventilation supplémentaires peuvent figurer dans le texte,,
