--- translated_by_canada/DisbursementChannel.csv
@@ -1,5 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Money is disbursed through central Ministry of Finance or Treasury,Les fonds sont déboursés par l'entremise du ministère des Finances ou du Trésor central,Money is disbursed through central Ministry of Finance or Treasury,Les fonds sont déboursés par l'entremise du ministère des Finances ou du Trésor central,,
-2,Money is disbursed directly to the implementing institution and managed through a separate bank account,Les fonds sont versés directement à l'organisation responsable de la mise en œuvre et gérés au moyen d'un compte bancaire distinct,Money is disbursed directly to the implementing institution and managed through a separate bank account,Les fonds sont versés directement à l'organisation responsable de la mise en œuvre et gérés au moyen d'un compte bancaire distinct,,
-3,"Aid in kind: Donors utilise third party agencies, e.g. NGOs or management companies","Aide en nature : Les donateurs utilisent des organisations tierces, p. ex. des ONG ou des sociétés de gestion","Aid in kind: Donors utilise third party agencies, e.g. NGOs or management companies","Aide en nature : Les donateurs utilisent des organisations tierces, p. ex. des ONG ou des sociétés de gestion",,
-4,Aid in kind: Donors manage funds themselves,Aide en nature : Les donateurs gèrent eux-mêmes les fonds,Aid in kind: Donors manage funds themselves,Aide en nature : Les donateurs gèrent eux-mêmes les fonds,,
