--- translated_by_canada/DocumentCategory.csv
@@ -1,31 +1,0 @@
-code,name,name (FR),description,description (FR),category,category (FR),url
-A01,Pre- and post-project impact appraisal,Évaluation de l'incidence avant et après le projet,,,A,A,
-A02,Objectives / Purpose of activity,Objectifs/raison d'être de l'activité,,,A,A,
-A03,Intended ultimate beneficiaries,Bénéficiaires finaux visés,,,A,A,
-A04,Conditions,Conditions,,,A,A,
-A05,Budget,Budget,,,A,A,
-A06,Summary information about contract,Renseignements sommaires sur le contrat,,,A,A,
-A07,Review of project performance and evaluation,Examen du rendement du projet et évaluation,,,A,A,
-A08,"Results, outcomes and outputs",Résultats et extrants,,,A,A,
-A09,Memorandum of understanding (If agreed by all parties),Protocole d'entente (s'il y a entente entre toutes les parties),,,A,A,
-A10,Tender,Appel d’offres,,,A,A,
-A11,Contract,Contrat,,,A,A,
-A12,Activity web page,Page Web sur l'activité,,,A,A,
-B01,Annual report,Rapport annuel,,,B,B,
-B02,Institutional Strategy paper,Stratégie organisationnelle,,,B,B,
-B03,Country strategy paper,Stratégie-pays,,,B,B,
-B04,Aid Allocation Policy,Politique sur la répartition de l'aide,,,B,B,
-B05,Procurement Policy and Procedure,Politique et procédure d'approvisionnement,,,B,B,
-B06,Institutional Audit Report,Rapport de vérification de l'organisation,,,B,B,
-B07,Country Audit Report,Rapport de vérification du pays,,,B,B,
-B08,Exclusions Policy,Politique sur les exclusions,,,B,B,
-B09,Institutional Evaluation Report,Rapport d'évaluation de l'organisation,,,B,B,
-B10,Country Evaluation Report,Rapport d'évaluation nationale,,,B,B,
-B11,Sector strategy,Stratégie sectorielle,,,B,B,
-B12,Thematic strategy,Stratégie thématique,,,B,B,
-B13,Country-level Memorandum of Understanding,Protocole d'entente national,,,B,B,
-B14,Evaluations policy,Politique sur les évaluations,,,B,B,
-B15,General Terms and Conditions,Modalités générales,,,B,B,
-B16,Organisation web page,Page Web sur l'organisation,,,B,B,
-B17,Country/Region web page,Page Web sur le pays/la région,,,B,B,
-B18,Sector web page,Page Web sur le secteur,,,B,B,
