--- translated_by_canada/GazetteerAgency.csv
@@ -1,4 +1,0 @@
-code,name,name (FR),description,category,url
-1,Geonames.org,Geonames.org,,,
-2,National Geospatial-Intelligence Agency,National Geospatial-Intelligence Agency,,,
-3,Open Street Map,Open Street Map,,,
