--- translated_by_canada/GeographicalPrecision.csv
@@ -1,10 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Exact location,Lieu précis,"The coordinates corresponds to an exact location, such as a populated place or a hill. The code is also used for locations that join a location which is a line (such as a road or railroad). Lines are not coded only the points that connect lines. All points that are mentioned in the source are coded.","Les coordonnées correspondent à un lieu précis, comme un endroit peuplé ou une colline. Le code est aussi utilisé pour les lieux qui rejoignent un autre endroit et qui est linéaire (comme une route ou une voie ferrée). Les lignes ne sont pas codées, seulement les points qui les forment. Tous les points mentionnés dans la source sont codés.",,
-2,Near exact location,Près d'un lieu précis,"The location is mentioned in the source as being ""near"", in the ""area"" of, or up to 25 km away from an exact location. The coordinates refer to that adjacent, exact, location.","Le lieu est mentionné dans la source comme étant « près », dans le « secteur » ou se trouvant à jusqu'à 25 km d'un lieu précis. Les coordonnées renvoient à ce lieu précis adjacent.",,
-3,Second order administrative division,Division administrative de second niveau,"The location is, or lies in, a second order administrative division (ADM2), such as a district, municipality or commune","Le lieu est ou se trouve dans une division administrative de second niveau (ADM2), comme un district, une municipalité ou une commune.",,
-4,First order administrative division,Division administrative de premier niveau,"The location is, or lies in, a first order administrative division (ADM1), such as a province, state or governorate.","Le lien est ou se trouve dans une division administrative de premier niveau (ADM1), comme une province, un État ou un gouvernorat.",,
-5,Estimated coordinates,Coordonnées estimées,"The location can only be related to estimated coordinates, such as when a location lies between populated places","On ne peut désigner le lieu qu'au moyen de coordonnées estimées, par exemple pour un lieu situé entre deux endroits peuplés.",,
-6,Independent political entity,Entité politique autonome,"The location can only be related to an independent political entity, meaning the pair of coordinates that represent a country.","On ne peut désigner le lieu que par rapport à une entité politique autonome, ce qui signifie la paire de coordonnées représentant un pays.",,
-7,Unclear - capital,Imprécis - Capitale,"Unclear. The capital is assumed to be one of two possible locations. (The other option is the country level, with precision 9.)","Imprécis. On suppose que la capitale est l'un de deux lieux possibles. (L'autre possibilité est le niveau du pays, avec la précision 9.)",,
-8,Local or national capital,Capitale locale ou nationale,"The location is estimated to be a seat of an administrative division (local capital) or the national capital. If aid goes to Luanda without further specification on the location, and there is an ADM1 and a capital called Luanda, then code the coordinates of the capital with precision 8. If it is not spelled out that aid goes to the capital  ","On estime que le lieu est le siège d'une division administrative (capitale locale) ou la capitale nationale. Si de l'aide est dirigée vers Luanda sans autre précision sur le lieu et qu'il existe une ADM1 et une capitale nommée Luanda, les coordonnées de la capitale doivent être inscrites avec la précision 8. En l'absence d'indication que l'aide va à la capitale",,
-9,Unclear - country,Imprécis - Pays,"Unclear. The locations is estimated to be the country level (often paired with the capital, with precision 7)","Imprécis. On estime que lieu est un pays (accompagne souvent la capitale, avec la précision 7)",,
