--- translated_by_canada/HumanitarianScopeType.csv
@@ -1,3 +1,0 @@
-code,name,name (FR),description,category,url
-1,Emergency,Urgence,,,
-2,Appeal,Appel,,,
