--- translated_by_canada/HumanitarianScopeVocabulary.csv
@@ -1,4 +1,0 @@
-code,name,name (FR),description,category,url
-1-2,Glide,Glide,,,http://glidenumber.net/glide/public/search/search.jsp
-2-1,Humanitarian Plan,Plan humanitaire,,,http://ftsarchive.unocha.org/docs/IATICodelist_HS2-1.csv
-99,Reporting Organisation,Organisation déclarante,,,
