--- translated_by_canada/LoanRepaymentPeriod.csv
@@ -1,5 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Annual,Annuel,The loan has 1 repayment per year,Le prêt a 1 remboursement par an,,
-2,Semi-annual,Semi-annuel,The loan has 2 repayments per year,Le prêt a 2 remboursements par an,,
-4,Quarterly,Trimestriel,The loan has 4 repayments per year,Le prêt a 4 remboursements par an,,
-12,Monthly,Mensuel,The loan has 12 monthly repayments per year,Le prêt a 12 remboursements par an,,
