--- translated_by_canada/LoanRepaymentType.csv
@@ -1,5 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Equal Principal Payments (EPP),Remboursements en tranches égales du principal,As required and defined by CRS++ reporting format Column 44,Tel que requis et défini dans le gabarit CRS++ colonne 44,,
-2,Annuity,Rente,As required and defined by CRS++ reporting format Column 44,Tel que requis et défini dans le gabarit CRS++ colonne 44,,
-3,Lump sum,Montant forfaitaire,As required and defined by CRS++ reporting format Column 44,Tel que requis et défini dans le gabarit CRS++ colonne 44,,
-5,Other,Autre,As required and defined by CRS++ reporting format Column 44,Tel que requis et défini dans le gabarit CRS++ colonne 44,,
