--- translated_by_canada/OrganisationRole.csv
@@ -1,5 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Funding,Financement,The government or organisation which provides funds to the activity.,Le gouvernement ou l'organisation qui finance l'activité.,,
-2,Accountable,Responsabilité,An organisation responsible for oversight of the activity and its outcomes,Une organisation responsable de la supervision de l'activité et de ses résultats,,
-3,Extending,Octroi,An organisation that manages the budget and direction of an activity on behalf of the funding organisation,Une organisation qui gère le budget et l'orientation d'une activité au nom de l'organisation offrant le financement,,
-4,Implementing,Mise en œuvre,The organisation that physically carries out the activity or intervention.,L'organisation qui exécute physiquement l'activité ou l'intervention.,,
