--- translated_by_canada/PolicyMarker.csv
@@ -1,10 +1,0 @@
-code,name,name (FR),description,category,url
-1,Gender Equality,Egalité des genres,,,
-2,Aid to Environment,Aide à l'environnement,,,
-3,Participatory Development/Good Governance,Développement participatif/Bonne gouvernance,,,
-4,Trade Development,Développement du commerce,,,
-5,Aid Targeting the Objectives of the Convention on Biological Diversity,Aide axée sur les objectifs de la Convention des Nations Unies sur la diversité biologique,,,
-6,Aid Targeting the Objectives of the Framework Convention on Climate Change - Mitigation,Aide axée sur les objectifs de la Convention-cadre des Nations Unies sur les changements climatiques - Atténuation,,,
-7,Aid Targeting the Objectives of the Framework Convention on Climate Change - Adaptation,Aide axée sur les objectifs de la Convention-cadre des Nations Unies sur les changements climatiques - Adaptation,,,
-8,Aid Targeting the Objectives of the Convention to Combat Desertification,Aide axée sur les objectifs de la Convention des Nations Unies sur la lutte contre la désertification,,,
-9,"Reproductive, Maternal, Newborn and Child Health (RMNCH)","Santé génésique, maternelle, néonatale et infantile",,,
