--- translated_by_canada/PolicyMarkerVocabulary.csv
@@ -1,3 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,OECD DAC CRS,OCDE CAD CRS,"The policy marker is an OECD DAC CRS policy marker, Reported in columns 20-23, 28-31 and 54 of CRS++ reporting format.","Le marqueur de politique est un marqueur de politique OCDE CAD SNPC, indiqué dans les colonnes 20-23, 28-31 et 54 du format de préparation de déclaration SNPC++.",,
-99,Reporting Organisation,Organisation déclarante,The policy marker is one that is defined and tracked by the reporting organisation,Le marqueur de politique est un marqueur défini et suivi par l'organisation déclarante,,
