--- translated_by_canada/ResultType.csv
@@ -1,5 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Output,Extrant,"Results of the activity that came about as a direct effect of your work and specific, what is done, and what communities are reached. For example, X number of individuals.","Résultats de l'activité découlant directement de votre travail et précisions connexes (ce qui est fait, les communautés rejointes). Par exemple, X nombre de personnes.",,
-2,Outcome,Effet,Results of the activity that produce an effect on the overall communities or issues you serve. For example lower rate of infection after a vaccination programme.,"Les résultats de l'activité qui produisent un effet sur les communautés générales ou les grands enjeux sur lesquels vous intervenez. Par exemple, taux d'infection réduit après une campagne de vaccination.",,
-3,Impact,Impact,"The long term effects of the outcomes, that lead to larger, over arching results, such as improved life-expectancy.","Les incidences à long terme des effets qui mènent à des résultats de niveau supérieur et de grande envergure, comme une amélioration de l'espérance de vie.",,
-9,Other,Autre,"Another type of result, not specified above.",Autre type de résultat non décrit ci-dessus.,,
