--- translated_by_canada/ResultVocabulary.csv
@@ -1,2 +1,0 @@
-code,name,name (FR),description,category,url
-99,Reporting Organisation,Organisation déclarante,,,
