--- translated_by_canada/SectorVocabulary.csv
@@ -1,13 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,OECD DAC CRS Purpose Codes (5 digit),Codes-secteurs OCDE CAD SNPC (5 chiffres),The sector reported corresponds to an OECD DAC CRS 5-digit purpose code,Le secteur déclaré correspond à un code de secteur OCDE CAD SNPC à cinq chiffres,,http://www.oecd.org/dac/stats/dacandcrscodelists.htm
-2,OECD DAC CRS Purpose Codes (3 digit),Codes-secteurs OCDE CAD SNPC (3 chiffres),The sector reported corresponds to an OECD DAC CRS 3-digit purpose code,Le secteur déclaré correspond à un code de secteur OCDE CAD SNPC à trois chiffres,,http://www.oecd.org/dac/stats/dacandcrscodelists.htm
-3,Classification of the Functions of Government (UN),Classification des fonctions gouvernementales (ONU),The sector reported corresponds to the UN Classification of the Functions of Government (CoFoG),Le secteur déclaré correspond à la Classification des fonctions gouvernementales (COFOG) de l'ONU,,http://unstats.un.org/unsd/cr/registry/regcst.asp?Cl=4
-4,Statistical classification of economic activities in the European Community,Classification statistique sur les activités économiques dans l'Union européenne,The sector reported corresponds to the statistical classifications of economic activities in the European Community,Le secteur déclaré correspond aux classifications statistiques des activités économiques dans l'Union européenne,,http://ec.europa.eu/eurostat/ramon/nomenclatures/index.cfm?TargetUrl=LST_NOM_DTL&StrNom=NACE_REV2&StrLanguageCode=EN
-5,National Taxonomy for Exempt Entities (USA),Taxonomie nationale des entités exemptes (États-Unis),The sector reported corresponds to the National Taxonomy for Exempt Entities (NTEE) - USA,Le secteur déclaré correspond à la taxonomie nationale des entités exonérées (NTEE) - États-Unis,,http://nccs.urban.org/classification/NTEE.cfm
-6,AidData,AidData,The sector reported corresponds to AidData classifications,Le secteur déclaré  correspond aux classifications AidData,,
-7,SDG Goal,ODD (objectif),A value from the top-level list of UN sustainable development goals (SDGs) (e.g. â€˜1â€™),Valeur tirée de la liste des principaux objectifs de développement durable (ODD) de l'ONU (p. ex. '1'),,https://sustainabledevelopment.un.org/?menu=1300
-8,SDG Target,ODD (cible),A value from the second-level list of UN sustainable development goals (SDGs) (e.g. â€˜1.1â€™),Valeur tirée de la liste des sous-objectifs de développement durable (ODD) de l'ONU (p. ex. '1.1'),,http://unstats.un.org/sdgs/indicators/indicators-list/
-9,SDG Indicator,ODD (indicateur),A value from the second-level list of UN sustainable development (SDG) indicators,Valeur tirée de la liste des indicateurs de développement durable (ODD) de deuxième niveau de l'ONU,,http://unstats.un.org/sdgs/indicators/indicators-list/
-10,Humanitarian Global Clusters (Inter-Agency Standing Committee),Groupes sectoriels d'aide humanitaire (Comité permanent interorganisations),The sector reported corresponds to an Inter-Agency Standard Committee Humanitarian Global Cluster code,Le secteur déclaré correspond à un code de groupe sectoriel du Comité permanent interorganisations,,https://data.humdata.org/dataset/standard-cluster-names-and-codes/resource/24b50d16-897b-46f9-848b-7fce08ac599b
-99,Reporting Organisation,Organisation déclarante,The sector reported corresponds to a sector vocabulary maintained by the reporting organisation for this activity,Le secteur déclaré correspond à un vocabulaire sectoriel maintenu par l'organisation déclarante aux fins de l'activité,,
-98,Reporting Organisation 2,Organisation déclarante 2,The sector reported corresponds to a sector vocabulary maintained by the reporting organisation for this activity (if they are referencing more than one),Le secteur déclaré correspond à un vocabulaire sectoriel maintenu par l'organisation déclarante aux fins de l'activité (si l'organisation en indique plus d'un),,
