--- translated_by_canada/TransactionType.csv
@@ -1,14 +1,0 @@
-code,name,name (FR),description,description (FR),category,url
-1,Incoming Funds,Fonds entrants,"Funds recieved for use on the activity, which can be from an external or internal source.",Fonds reçus pour être utilisés dans le cadre de l'activité et qui peuvent provenir d'une source externe ou interne.,,
-2,Outgoing Commitment,Engagement sortant,"A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, for specific purposes, for the benefit of the recipient.","Une obligation ferme et écrite d'un donateur ou d'un fournisseur d'offrir au bénéficiaire un financement d'un montant établi, conformément à des modalités précises, dans un but explicite.",,
-3,Disbursement,Décaissement,"Outgoing funds that are placed at the disposal of a recipient government or organisation, or funds transferred between two separately reported activities.",Fonds sortants mis à la disposition d'un gouvernement ou d'une organisation bénéficiaire ou fonds transférés entre deux activités faisant l'objet de déclarations distinctes.,,
-4,Expenditure,Dépenses,Outgoing funds that are spent on goods and services for the activity.,Fonds sortants dépensés pour acquérir des biens et services pour l'activité.,,
-5,Interest Payment,Paiements d’intérêts,"The actual amount of interest paid on a loan or line of credit, including fees.","Le montant réel des intérêts payés en lien avec un prêt ou une marge de crédit, y compris les frais.",,
-6,Loan Repayment,Remboursement du prêt,"The actual amount of principal (amortisation) repaid, including any arrears.","Le montant réel du capital remboursé (amortissement), y compris les arriérés, le cas échéant.",,
-7,Reimbursement,Remboursement,"A type of disbursement that covers funds that have already been spent by the recipient, as agreed in the terms of the grant or loan","Un type de décaissement qui couvre des fonds qui ont déjà été dépensés par le bénéficiaire, conformément aux modalités de la subvention ou du prêt",,
-8,Purchase of Equity,Achat d'équité,Outgoing funds that are used to purchase equity in a business,Fonds sortants qui sont utilisés pour acquérir de l'équité dans une entreprise,,
-9,Sale of Equity,Vente d'équité,Incoming funds from the sale of equity.,Fonds reçus en lien avec la vente d'équité,,
-10,Credit Guarantee,Caution,A commitment made by a funding organisation to underwrite a loan or line of credit entered into by a third party.,Engagement pris par une organisation de financement pour garantir un prêt ou une marge de crédit contractés par une tierce partie.,,
-11,Incoming Commitment,Engagement entrant,"A firm, written obligation from a donor or provider to provide a specified amount of funds, under particular terms and conditions, reported by a recipient for this activity.","Une obligation ferme et écrite d'un donateur ou d'un fournisseur d'offrir un financement d'un montant établi, conformément à des modalités précises et déclarée par le bénéficiaire en lien avec l'activité.",,
-12,Outgoing Pledge,Promesse sortante,"Indicative, non-binding advice of an intended outgoing commitment.",Avis indicatif et sans obligation indiquant une intention de prendre un engagement sortant.,,
-13,Incoming Pledge,Promesse entrante,"Indicative, non-binding advice of an intended incoming commitment.",Avis indicatif et sans obligation indiquant une intention de prendre un engagement entrant.,,
